// packages/config/src/generated/writable.ts
export const WRITABLE = {
  "store": {
    "create": null,
    "update": {
      "properties": {
        "name": {
          "type": "string"
        },
        "mail_from_address": {
          "type": "string"
        },
        "customer_support_email": {
          "type": "string",
          "nullable": true
        },
        "new_order_notifications_email": {
          "type": "string",
          "nullable": true
        },
        "preferred_admin_locale": {
          "type": "string",
          "nullable": true
        },
        "preferred_timezone": {
          "type": "string"
        },
        "preferred_weight_unit": {
          "type": "string"
        },
        "preferred_unit_system": {
          "type": "string"
        },
        "preferred_storefront_access": {
          "type": "string",
          "enum": [
            "public",
            "prices_hidden",
            "login_required"
          ]
        },
        "preferred_storefront_url": {
          "type": "string",
          "nullable": true
        },
        "preferred_guest_checkout": {
          "type": "boolean"
        },
        "preferred_always_include_confirm_step": {
          "type": "boolean"
        },
        "preferred_company_field_enabled": {
          "type": "boolean"
        },
        "preferred_address_requires_company": {
          "type": "boolean"
        },
        "preferred_address_requires_phone": {
          "type": "boolean"
        },
        "preferred_capture_method": {
          "type": "string",
          "enum": [
            "checkout",
            "on_dispatch",
            "manual"
          ]
        },
        "preferred_auto_capture": {
          "type": "boolean",
          "deprecated": true
        },
        "preferred_auto_capture_on_dispatch": {
          "type": "boolean"
        },
        "preferred_track_inventory_levels": {
          "type": "boolean"
        },
        "preferred_stock_reservations_enabled": {
          "type": "boolean"
        },
        "preferred_low_stock_threshold": {
          "type": "integer"
        },
        "preferred_tax_using_ship_address": {
          "type": "boolean"
        },
        "preferred_track_price_history": {
          "type": "boolean"
        },
        "preferred_show_products_without_price": {
          "type": "boolean"
        },
        "preferred_disable_sku_validation": {
          "type": "boolean"
        },
        "preferred_order_routing_strategy": {
          "type": "string"
        },
        "preferred_pricing_provider": {
          "type": "string"
        },
        "preferred_inventory_provider": {
          "type": "string"
        },
        "preferred_pricing_provider_failure_policy": {
          "type": "string",
          "enum": [
            "fallback",
            "strict"
          ]
        },
        "preferred_inventory_provider_failure_policy": {
          "type": "string",
          "enum": [
            "fallback",
            "strict"
          ]
        },
        "preferred_payout_provider": {
          "type": "string",
          "nullable": true
        },
        "preferred_default_payouts_schedule_interval": {
          "type": "string",
          "enum": [
            "daily",
            "weekly",
            "biweekly",
            "monthly",
            "manual"
          ]
        },
        "preferred_default_minimum_payout_amount": {
          "type": "number"
        },
        "preferred_auto_approve_sellers": {
          "type": "boolean"
        },
        "preferred_auto_approve_seller_products": {
          "type": "boolean"
        },
        "preferred_send_seller_transactional_emails": {
          "type": "boolean"
        },
        "preferred_send_consumer_transactional_emails": {
          "type": "boolean"
        },
        "preferred_default_commission_tax_rate": {
          "type": "number"
        },
        "preferred_document_number_format": {
          "type": "string",
          "enum": [
            "sequential",
            "random"
          ]
        },
        "preferred_order_number_prefix": {
          "type": "string"
        },
        "preferred_order_number_suffix": {
          "type": "string"
        },
        "preferred_order_number_sequence_start": {
          "type": "integer"
        },
        "mailer_logo": {
          "type": "string"
        }
      },
      "required": []
    }
  },
// ...
} as const